BO: Update wording form.html.twig
If you mean the item number (or article number), why don't you name it correctly? This is neither a *"reference code"* nor is it *"Internal"*. Just the other way around: The item number (what you call *reference*) is displayed on the shop front in product details. So nothing is internal - this field's content **is** the item number for the merchant **and** the customer!