FO: Fixes wrong "InStock" microdata for out of stock products
In Classic, there is a wrong schema of "InStock" for all products, regardless if the product is actually out of stock. If the product is out of stock, the right schema should be "OutofStock". If the product is Out of stock but orders are alowed, the schema should be "PreOrder".