Fix secure_key missing from invoice links for guests. Original fix by @BitcoinMitchell.
Issue: Links to invoices for guests on the order confirmation and guest tracking pages did not have their secure_key set, prompting them to log in.
Cause: $context->cookie->is_guest does not appear to be set in those contexts. Also, the generated link contains a pre-encoded ampersand which should not be present.

Alternative fixes (may be more "correct"):
- [WORKS] Use $context->customer->id_guest in HistoryController.php:87 (Works in both order confirmation and guest tracking pages.)
- [DOESN'T WORK] Use $context->customer->is_guest (Set properly in order confirmation page but not in guest tracking if the viewer does not have cookies from the previous session.)