Fixed "Add new alias" button
The fix is about "add new alias" button in Shop Parameters > Search which wasn't behaving properly, as well explained in issue PrestaShop#21274 by @ngodefroy.
It is a tiny addition to the if statement in AdminSearchConfController::initPageHeaderToolbar().
On the first click to "Search" BO page $this->display == null . When you add a new alias and save instead $this->display == 'list'. Despite the fact I would prefer $this->display to be equal to 'list' at the first side (first click to search); I decided to just allow both situations. This shouldn't impact, in other ways, alias listing/editing/creation process.