Update AdminImportController.php
The problem depends on the fact that when importing store contacts, the "PrestaShopLogger->object_type property" is set with "Store contacts" which does not pass the validation of the Validae::isValidObjectClassName() method.
So in this case I thought of valorising the $import_type variable with "StoreContacts" which passes the validation and the error no longer occurs.